updated the docker script to check if container exists.

+++ week08/docker.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
-docker run -d --rm -h banana --name banana -v $(pwd):/ctf/work --cap-add=SYS_PTRACE plsiamlegit/6447-ubuntu:pwndocker
+if [[ "$(docker image inspect plsiamlegit/6447-ubuntu 2> /dev/null)" == "" ]]; then
+    docker run -d --rm -h banana --name banana -v $(pwd):/ctf/work --cap-add=SYS_PTRACE plsiamlegit/6447-ubuntu:pwndocker
+fi
 
 docker exec -it banana /bin/bash
